requirements.txt
Django==2.2.0
graphql_core==2.1.0
graphene==2.1.5
graphene_django==2.2.0
django_graphql_jwt==0.2.1
pytz==2018.9
graphene-file-upload=1.2.2
django-filter==2.1.0
django-filters==0.2.1